Moved `encode` and `decode` from `model` to `client` module
