+++ modules/model/src/main/kotlin/jp/co/soramitsu/iroha2/util.kt
@@ -1,5 +1,9 @@
 package jp.co.soramitsu.iroha2
 
+import io.emeraldpay.polkaj.scale.ScaleCodecWriter
+import io.emeraldpay.polkaj.scale.reader.UInt128Reader
+import java.math.BigInteger
+
 // Copied from Google Guava library(com.google.common.collect.Maps)
 
 /**
@@ -29,5 +33,24 @@ fun mapCapacity(expectedSize: Int): Int {
         expectedSize < 3 -> expectedSize + 1
         expectedSize < MAX_POWER_OF_TWO -> (expectedSize.toFloat() / 0.75f + 1.0f).toInt()
         else -> Int.MAX_VALUE
+    }
+}
+
+fun writeUint64(writer: ScaleCodecWriter, value: Long) {
+    require(value >= 0) { "Negative values are not supported: $value" }
+    val array = BigInteger.valueOf(value).toByteArray()
+    var pos = 0
+    if (array[0].toInt() == 0) {
+        ++pos
+    }
+
+    val len = array.size - pos
+    if (len > 8) {
+        throw IllegalArgumentException("Value is to big for 64 bits. Has: " + len * 8 + " bits")
+    } else {
+        val encoded = ByteArray(8)
+        System.arraycopy(array, pos, encoded, encoded.size - len, len)
+        UInt128Reader.reverse(encoded)
+        writer.directWrite(encoded, 0, 8)
     }
 }
